asset_pipeline/generators/gemini_image.py
import os
from typing import Any, Dict, List
from google import genai
from google.genai import types


class GeminiImageGenerator:
    """Wrapper for Gemini image generation models."""
    
    def __init__(self, api_key: str = None):
        self.api_key = api_key or os.getenv("GOOGLE_API_KEY")
        if not self.api_key:
            raise ValueError("GOOGLE_API_KEY environment variable required")
        
        self.client = genai.Client(api_key=self.api_key)
    
    def generate_images(
        self,
        prompt: str,
        config: Dict[str, Any],
        safety_settings: List[Dict[str, str]]
    ) -> List[Dict[str, Any]]:
        """Generate images using Gemini image models."""
        
        # Convert safety settings
        safety_config = []
        for setting in safety_settings:
            safety_config.append(
                types.SafetySetting(
                    category=getattr(types.HarmCategory, setting["category"]),
                    threshold=getattr(types.HarmBlockThreshold, setting["threshold"])
                )
            )
        
        # For Gemini image models, use text generation with image output
        model_name = config.get("model", "gemini-3-pro-image-preview")
        
        response = self.client.models.generate_content(
            model=model_name,
            contents=prompt,
            config=types.GenerateContentConfig(
                safety_settings=safety_config,
                # Add image generation specific config if available
            )
        )
        
        # Mock response format - adjust based on actual API
        results = []
        for i in range(config.get("numberOfImages", 1)):
            results.append({
                "id": f"gemini_{i}",
                "url": f"mock_gemini_url_{i}",  # Replace with actual image URL
                "metadata": {
                    "model": model_name,
                }
            })
        
        return results