asset_pipeline/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asset-pipeline"
version = "0.1.0"
description = "Multi-Asset Image Generation Pipeline for Game Assets"
authors = [
    {name = "AI Engineer", email = "ai@example.com"}
]
dependencies = [
    "pydantic>=2.0.0",
    "langgraph>=0.1.0",
    "google-genai>=0.1.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "pytest>=7.0.0",
]

[project.scripts]
asset-pipeline = "asset_pipeline.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["asset_pipeline*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]